move Device trait from mod memory -> devices
this commit also moves memory::util to crate::util
