+++ src/devices/uart/device.rs
@@ -6,8 +6,8 @@ use std::time::Duration;
 use crossbeam_channel::{self as chan, select};
 use log::*;
 
-use crate::devices::vic::Interrupt;
-use crate::memory::{Device, MemException::*, MemResult, Memory, Probe};
+use crate::devices::{vic::Interrupt, Device, Probe};
+use crate::memory::{MemException::*, MemResult, Memory};
 
 /// Aggregate type to configure which Interrupts should be generated by the UART
 #[derive(Debug)]
