+++ src/memory/mod.rs
@@ -1,27 +1,10 @@
 pub mod armv4t_adaptor;
-pub mod util;
 
 mod access;
 mod exception;
 
 pub use access::{MemAccess, MemAccessKind, MemAccessVal};
 pub use exception::{MemException, MemResult};
-
-/// Implemented by all emulated devices.
-/// Provides a way to traverse and query the device tree.
-pub trait Device {
-    /// The name of the emulated device.
-    fn kind(&self) -> &'static str;
-
-    /// A descriptive label for a particular instance of the device
-    /// (if applicable).
-    fn label(&self) -> Option<&str> {
-        None
-    }
-
-    /// Query what devices exist at a particular memory offset.
-    fn probe(&self, offset: u32) -> Probe<'_>;
-}
 
 /// Common memory trait used throughout the emulator.
 ///
@@ -70,41 +53,6 @@ pub trait Memory {
     }
 }
 
-/// A link in a chain of devices corresponding to a particular memory offset.
-pub enum Probe<'a> {
-    /// Branch node representing a device.
-    Device {
-        device: &'a dyn Device,
-        next: Box<Probe<'a>>,
-    },
-    /// Leaf node representing a register.
-    Register(&'a str),
-    /// Unmapped memory.
-    Unmapped,
-}
-
-impl<'a> std::fmt::Display for Probe<'a> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
-        match self {
-            Probe::Device { device, next } => {
-                match device.label() {
-                    Some(label) => write!(f, "{}:{}", device.kind(), label)?,
-                    None => write!(f, "{}", device.kind())?,
-                };
-
-                match &**next {
-                    Probe::Unmapped => {}
-                    next => write!(f, " > {}", next)?,
-                }
-            }
-            Probe::Register(name) => write!(f, "{}", name)?,
-            Probe::Unmapped => write!(f, "<unmapped>")?,
-        }
-
-        Ok(())
-    }
-}
-
 macro_rules! impl_memfwd {
     ($type:ty) => {
         impl Memory for $type {
@@ -135,27 +83,5 @@ macro_rules! impl_memfwd {
     };
 }
 
-macro_rules! impl_devfwd {
-    ($type:ty) => {
-        impl Device for $type {
-            fn kind(&self) -> &'static str {
-                (**self).kind()
-            }
-
-            fn label(&self) -> Option<&str> {
-                (**self).label()
-            }
-
-            fn probe(&self, offset: u32) -> Probe<'_> {
-                (**self).probe(offset)
-            }
-        }
-    };
-}
-
 impl_memfwd!(Box<dyn Memory>);
 impl_memfwd!(&mut dyn Memory);
-
-impl_devfwd!(Box<dyn Device>);
-impl_devfwd!(&dyn Device);
-impl_devfwd!(&mut dyn Device);
